requirements.txt
typer>=0.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0
pydantic>=2.0.0
pyyaml>=6.0
gitpython>=3.1.0
litellm>=1.0.0
rich>=13.0.0
apscheduler>=3.10.0
psycopg2-binary>=2.9
tenacity>=8.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
